Move `Roll.UnappliedMigrations` to its own file
And rename the test file accordingly.
